+++ packages/graphql-api/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @crux/graphql-api
+# @gokceno/crux-graphql-api
 
 ## 0.0.1
 
@@ -6,4 +6,4 @@
 
 - Init Changesets
 - Updated dependencies
-  - @crux/graphql-api-resolvers@0.0.1
+  - @gokceno/crux-graphql-api-resolvers@0.0.1