+++ packages/graphql-schema/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @crux/graphql-schema
+# @gokceno/crux-graphql-schema
 
 ## 0.0.1
 
@@ -6,5 +6,5 @@
 
 - Init Changesets
 - Updated dependencies
-  - @crux/graphql-api@0.0.1
-  - @crux/graphql-api-resolvers@0.0.1
+  - @gokceno/crux-graphql-api@0.0.1
+  - @gokceno/crux-graphql-api-resolvers@0.0.1